+++ src/countertimer.rs
@@ -32,7 +32,8 @@ impl SilenceCounter {
         chunksize: usize,
     ) -> SilenceCounter {
         let silence_threshold = PrcFmt::new(10.0).powf(silence_threshold_db / 20.0);
-        let silence_limit_nbr = (silence_timeout * ((samplerate / chunksize) as PrcFmt)) as usize;
+        let silence_limit_nbr =
+            (silence_timeout * samplerate as PrcFmt / chunksize as PrcFmt).round() as usize;
         SilenceCounter {
             silence_threshold,
             silence_limit_nbr,
@@ -238,7 +239,7 @@ mod tests {
     #[test]
     fn silencecounter() {
         let mut counter = SilenceCounter::new(-40.0, 3.0, 48000, 1024);
-        let limit_nbr = 3 * (48000 / 1024);
+        let limit_nbr = (3.0f64 * 48000.0 / 1024.0).round() as usize;
         assert_eq!(counter.silence_limit_nbr, limit_nbr);
         assert_eq!(counter.silence_threshold, 0.01);
         for _ in 0..(2 * limit_nbr) {
@@ -254,6 +255,30 @@ mod tests {
             assert_eq!(state, ProcessingState::Paused);
         }
         for _ in 0..(2 * limit_nbr) {
+            let state = counter.update(0.1);
+            assert_eq!(state, ProcessingState::Running);
+        }
+    }
+
+    #[test]
+    fn silencecounter_largechunksize() {
+        let mut counter = SilenceCounter::new(-40.0, 1.0, 48000, 23000);
+        let limit_nbr = 2;
+        assert_eq!(counter.silence_limit_nbr, limit_nbr);
+        assert_eq!(counter.silence_threshold, 0.01);
+        for _ in 0..5 {
+            let state = counter.update(0.1);
+            assert_eq!(state, ProcessingState::Running);
+        }
+        for _ in 0..2 {
+            let state = counter.update(0.001);
+            assert_eq!(state, ProcessingState::Running);
+        }
+        for _ in 0..5 {
+            let state = counter.update(0.001);
+            assert_eq!(state, ProcessingState::Paused);
+        }
+        for _ in 0..5 {
             let state = counter.update(0.1);
             assert_eq!(state, ProcessingState::Running);
         }