+++ src/loudness.rs
@@ -237,7 +237,7 @@ impl Filter for Loudness {
     }
 }
 
-/// Validate a FFT convolution config.
+/// Validate a Loudness config.
 pub fn validate_config(conf: &config::LoudnessParameters) -> Res<()> {
     if conf.reference_level > 0.0 {
         return Err(config::ConfigError::new("Reference level must be less than 0").into());
@@ -251,6 +251,8 @@ pub fn validate_config(conf: &config::LoudnessParameters) -> Res<()> {
         return Err(config::ConfigError::new("High boost cannot be larger than 20").into());
     } else if conf.low_boost > 20.0 {
         return Err(config::ConfigError::new("Low boost cannot be larger than 20").into());
+    } else if conf.ramp_time < 0.0 {
+        return Err(config::ConfigError::new("Ramp time cannot be negative").into());
     }
     Ok(())
 }